tast-tests: save the screenshot file during the test
The screenshot used to be saved during the test and it's helpful to
check the screenshot if the test fails. Due to change in
crrev.com/c/1691882, the screenshot is removed when the test finished.
To have the option to keep the screenshot files after the test, an
additional parameter is added in the screenshot.GrabScreenshot function.

Currently, arc.ResizeActivity and arc.QuarterSizedWindowZooming will
save the screenshot in this patch. Recommended for other tests too.

BUG=b:167419204
TEST=tast -verbose run <DUT> arc.BlackFlash
TEST=tast -verbose run <DUT> arc.ResizeActivity
TEST=tast -verbose run <DUT> arc.QuarterSizedWindowZooming

+++ src/chromiumos/tast/local/bundles/cros/arc/screenshot/screenshot.go
@@ -73,17 +73,34 @@ func CountDiffPixels(imageA, imageB image.Image, threshold uint8) (int, error) {
 }
 
 // GrabScreenshot creates a screenshot and returns an image.Image.
-// The path of the image is generated ramdomly in /tmp.
-func GrabScreenshot(ctx context.Context, cr *chrome.Chrome) (image.Image, error) {
-	fd, err := ioutil.TempFile("", "screenshot")
-	if err != nil {
-		return nil, errors.Wrap(err, "error opening screenshot file")
+// If the path is nil, the path of the image is generated ramdomly in /tmp and the screenshot will be deleted when the test is finished.
+// If the path is assigned, the screenshot won't be deleted when the test is finished.
+func GrabScreenshot(ctx context.Context, cr *chrome.Chrome, path *string) (image.Image, error) {
+	var fd *os.File
+	var dstPath string
+	var err error
+	if path == nil {
+		fd, err = ioutil.TempFile("", "screenshot")
+		if err != nil {
+			return nil, errors.Wrap(err, "error opening screenshot file")
+		}
+		defer os.Remove(fd.Name())
+		defer fd.Close()
+		dstPath = fd.Name()
+	} else {
+		dstPath = *path
 	}
-	defer os.Remove(fd.Name())
-	defer fd.Close()
 
-	if err := screenshot.CaptureChrome(ctx, cr, fd.Name()); err != nil {
+	if err = screenshot.CaptureChrome(ctx, cr, dstPath); err != nil {
 		return nil, errors.Wrap(err, "failed to capture screenshot")
+	}
+
+	if path != nil {
+		fd, err = os.Open(*path)
+		if err != nil {
+			return nil, errors.Wrap(err, "error opening screenshot file")
+		}
+		defer fd.Close()
 	}
 
 	img, _, err := image.Decode(fd)
